+++ changelog.md
@@ -82,6 +82,10 @@ Pages
 - `/journals`
   - POST
   - GET
+- `/journals/categories/`
+  - GET
+- `/journals/categories/:category`
+  - GET
 - `/journals/:journalId`
   - PUT
   - DELETE
@@ -94,8 +98,6 @@ Pages
 - `/journals/:journalId/comments/:commentId`
   - PUT
   - DELETE
-- `/categories/`
-  - GET
 
 ## Changes
 
@@ -106,5 +108,6 @@ Pages
 - stub basic architecture
 - create basic api, GET `/`
 - create journels router and mount on `/journals/`
+- stub all api routes on journals router
 
 ### Misc